dts: gpucc: increase performance by upscaling gfx3d
If you need a more stable and faster performance in opening applications as well as lightness and smoothness in playing games then this commit will provide it

In here I don't make the frequency 19xxxxx become frequency 0 because frequency 0 will not run in the application / when the application is run it will remain silent in the position or maybe it will only run when deepsleep starts. I don't know about this but it's possible that the gpu won't run more than 20xxxxx in deepsleep

Reference blog: https://www.analog.com/en/analog-dialogue/articles/phase-locked-loop-pll-fundamentals.html

After I read a little from the source above, I tried it to implement it here.