tcp: decrease somaxconn for better ms
So yeah in the previous commit which says "increase somaxconn or about this line" if you make the number too big for example 1xxx/2xxx/3xxx/4xxx and so on then your internet load will be greatly reduced

Or in this sense. If the number is getting bigger then before the internet signal is sent to the route it will be accommodated first. If the shelter is at its maximum, the signal from the new route will be sent to our device

"I DO NOT RECOMMEND YOU CHANGE THE VALUE BIGGER"

But it all comes back to your stance. Because I've checked several times on certain websites about this somaxconn and the results I get are different "in terms of load already different"