defconfig: i think gov schedutil is bad for game
I tested it for a while, the schedutil governor consumes less power and it's like there is pressure exerted by this gov. In a few days I tested it in the game HI3 (honkai impact 3). When I changed the governor to a schedutil frame what happened to the game was even worse. Even though I have lowered the cpu voltage by 20mv but as I said. In the previous test, there were no very severe frames when playing the double x game. Even though I understand that the cellphone I use will still frame. But if there is a difference I will feel it right away. Because I tested this kernel for several days even days/months to create this stable kernel.

BUG KERNEL: Screen OF Death

As for how to solve this, I'm not sure how. but I think there's something wrong with my defconfig causing the SOD, maybe next time I'll check it again in the section input, touchscreen, and display panels. because temporarily SOD can be overcome by turning on AOD.

FIX: Turn ON The AOD

But I'm also not sure if this is the only thing that happens on my device or on all other devices, I don't know that. Also because my device is weak and has a lot of damage what happened to my device, so I don't know, the damage I mean here, such as a broken screen, the vibrator comes off the hook and others. (Because my device has been old and once fallen)